+++ packages/brick_offline_first_with_rest_build/test/offline_first_generator/test_one_to_many_association.dart
@@ -104,7 +104,7 @@ class OneToManyAssociationAdapter
         final id = s?.primaryKey ??
             await provider?.upsert<SqliteAssoc>(s, repository: repository);
         return await provider?.rawInsert(
-            'INSERT OR REPLACE INTO `_brick_OneToManyAssociation_assoc` (`l_OneToManyAssociation_brick_id`, `f_SqliteAssoc_brick_id`) VALUES (?, ?)',
+            'INSERT OR IGNORE INTO `_brick_OneToManyAssociation_assoc` (`l_OneToManyAssociation_brick_id`, `f_SqliteAssoc_brick_id`) VALUES (?, ?)',
             [instance.primaryKey, id]);
       }));
     }